+++ indispensable-asn1/src/commonMain/kotlin/at/asitplus/signum/indispensable/asn1/Asn1Elements.kt
@@ -5,17 +5,10 @@ package at.asitplus.signum.indispensable.asn1
 import at.asitplus.catching
 import at.asitplus.catchingUnwrapped
 import at.asitplus.signum.indispensable.asn1.Asn1Element.Tag.Template.Companion.withClass
-import at.asitplus.signum.indispensable.asn1.Asn1PrimitiveOctetString
 import at.asitplus.signum.indispensable.asn1.encoding.*
 import kotlinx.io.Buffer
 import kotlinx.io.Sink
 import kotlinx.io.readByteArray
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
 import kotlin.experimental.ExperimentalObjCName
 import kotlin.native.ObjCName
 
@@ -36,6 +29,13 @@ sealed class Asn1Element(
         return true
 
     }
+
+    /**
+     * The DER-encoded content bytes of this ASN.1 element.
+     * **Accessing this property on ASN.1 structures triggers encoding of the contained child nodes**, which may cause
+     * performance hits. Use with caution!
+     */
+    abstract val content: ByteArray
 
     companion object {
         /**
@@ -190,7 +190,7 @@ sealed class Asn1Element(
 
     @Throws(Asn1StructuralException::class)
     private inline fun <reified T> thisAs(): T =
-       (this as? T)
+        (this as? T)
             ?: throw Asn1StructuralException("${this::class.simpleName} cannot be reinterpreted as ${T::class.simpleName}.")
 
 
@@ -319,11 +319,62 @@ sealed class Asn1Element(
             val TIME_GENERALIZED = Tag(tagValue = BERTags.GENERALIZED_TIME.toULong(), constructed = false)
             val TIME_UTC = Tag(tagValue = BERTags.UTC_TIME.toULong(), constructed = false)
 
+            val entries: Iterable<Tag> by lazy {
+                listOf(
+                    SET,
+                    SEQUENCE,
+                    NULL,
+                    BOOL,
+                    INT,
+                    REAL,
+                    OID,
+                    ENUM,
+                    OCTET_STRING,
+                    BIT_STRING,
+                    STRING_UTF8,
+                    STRING_UNIVERSAL,
+                    STRING_IA5,
+                    STRING_BMP,
+                    STRING_T61,
+                    STRING_PRINTABLE,
+                    STRING_NUMERIC,
+                    STRING_VISIBLE,
+                    TIME_GENERALIZED,
+                    TIME_UTC
+                )
+            }
+
         }
 
         val tagClass by lazy {
             checkNotNull(TagClass.fromByte(encodedTag.first()).getOrNull()) {
                 "An Illegal Tag class has been found. This should be impossible!"
+            }
+        }
+
+        val name by lazy {
+            when (this) {
+                SET -> "SET"
+                SEQUENCE -> "SEQUENCE"
+                NULL -> "NULL"
+                BOOL -> "BOOLEAN"
+                INT -> "INTEGER"
+                REAL -> "REAL"
+                OID -> "OBJECT IDENTIFIER"
+                ENUM -> "ENUMERATED"
+                OCTET_STRING -> "OCTET STRING"
+                BIT_STRING -> "BIT STRING"
+                STRING_UTF8 -> "UTF8 STRING"
+                STRING_UNIVERSAL -> "UNIVERSAL STRING"
+                STRING_IA5 -> "IA5 STRING"
+                STRING_BMP -> "BMP STRING"
+                STRING_T61 -> "T61 STRING"
+                STRING_PRINTABLE -> "PRINTABLE STRING"
+                STRING_NUMERIC -> "NUMERIC STRING"
+                STRING_VISIBLE -> "VISIBLE STRING"
+                TIME_GENERALIZED -> "GENERALIZED TIME"
+                TIME_UTC -> "UTC TIME"
+                else -> null
             }
         }
 
@@ -333,7 +384,7 @@ sealed class Asn1Element(
 
         override fun toString(): String =
             "${tagClass.let { if (it == TagClass.UNIVERSAL) "" else it.name + " " }}${tagValue}${if (isConstructed) " CONSTRUCTED" else ""}" +
-                    (" (=${encodedTag.toHexString(HexFormat.UpperCase)})")
+                    (" (=${encodedTag.toHexString(HexFormat.UpperCase)})" + (name?.let { " ($it)" } ?: ""))
 
         /**
          * As per ITU-T X.680 8824-1 8.6
@@ -435,6 +486,10 @@ inline fun <reified T : Asn1Element> T.assertTag(tag: Asn1Element.Tag): T {
 @Throws(Asn1TagMismatchException::class)
 inline fun <reified T : Asn1Element> T.assertTag(tagNumber: ULong): T = assertTag(tag withNumber tagNumber)
 
+/**
+ * ASN.1 NULL object as constant
+ */
+val Asn1Null = Asn1Primitive(Asn1Element.Tag.NULL, byteArrayOf())
 
 /**
  * ASN.1 structure. Contains no data itself, but holds zero or more [children]
@@ -541,6 +596,13 @@ sealed class Asn1Structure(
 
         return true
     }
+
+    /**
+     * The DER-encoded content bytes of this ASN.1 element.
+     * **Accessing this property on ASN.1 structures triggers encoding of the contained child nodes**, which may cause
+     * performance hits. Use with caution!
+     */
+    override val content: ByteArray by lazy { children.fold(byteArrayOf()) { acc, asn1Element -> acc + asn1Element.derEncoded } }
 }
 
 /**
@@ -650,17 +712,6 @@ class Asn1CustomStructure private constructor(
         shouldBeSorted: Boolean = false
     ) : this(children, tag.toULong(), tagClass, sortChildren, shouldBeSorted)
 
-
-    /**
-     * Raw byte DER-encoded representation of this custom structure's children.
-     * This property is `null` **unless** the `CONSTRUCTED` flag of this structure's tag is overridden to `false`
-     */
-    val content: ByteArray? by lazy {
-        if (!tag.isConstructed)
-            children.fold(byteArrayOf()) { acc, asn1Element -> acc + asn1Element.derEncoded }
-        else null
-    }
-
     override fun toString() = "${tag.tagClass}" + super.toString()
 
     override fun prettyPrintHeader(indent: Int) =
@@ -669,7 +720,7 @@ class Asn1CustomStructure private constructor(
                 (if (!tag.isConstructed) " PRIMITIVE" else "") +
                 " (=${tag.encodedTag.toHexString(HexFormat.UpperCase)}), length=${length}" +
                 ", overallLength=${overallLength}" +
-                content?.let { " ${it.toHexString(HexFormat.UpperCase)}" }
+                " ${content.toHexString(HexFormat.UpperCase)}"
 
     companion object {
         /**
@@ -703,9 +754,6 @@ class Asn1CustomStructure private constructor(
 class Asn1EncapsulatingOctetString(children: List<Asn1Element>) :
     Asn1Structure(Tag.OCTET_STRING, children, sortChildren = false, shouldBeSorted = false),
     Asn1OctetString {
-    override val content: ByteArray by lazy {
-        children.fold(byteArrayOf()) { acc, asn1Element -> acc + asn1Element.derEncoded }
-    }
 
     override fun equals(other: Any?): Boolean {
         if (other is Asn1PrimitiveOctetString) return this.content contentEquals other.content
@@ -785,7 +833,7 @@ open class Asn1Primitive(
     /**
      * Raw data contained in this ASN.1 primitive in its encoded form. Requires decoding to interpret it
      */
-    val content: ByteArray
+    override val content: ByteArray
 ) : Asn1Element(tag) {
     init {
         if (tag.isConstructed) throw IllegalArgumentException("A primitive cannot have a CONSTRUCTED tag")
@@ -804,8 +852,32 @@ open class Asn1Primitive(
 
     override fun prettyPrintHeader(indent: Int) = (" " * indent) + "Primitive" + super.prettyPrintHeader(indent)
 
-    override fun contentToString() = content.toHexString(HexFormat.UpperCase)
-    override fun prettyPrintContents(indent: Int) = contentToString()
+    override fun contentToString() = catchingUnwrapped {
+        when (tag) {
+            Tag.NULL -> ""
+            Tag.BOOL -> decodeToBoolean().toString()
+            Tag.INT -> decodeToInt().toString()
+            Tag.REAL -> decodeToFloat().toString()
+            Tag.OID -> ObjectIdentifier.decodeFromAsn1ContentBytes(content).toString()
+            Tag.ENUM -> decodeToEnumOrdinal().toString()
+            Tag.OCTET_STRING -> content.toHexString(HexFormat.UpperCase)
+            Tag.BIT_STRING -> content.toHexString(HexFormat.UpperCase)
+            Tag.STRING_UTF8 -> decodeToString()
+            Tag.STRING_UNIVERSAL -> decodeToString()
+            Tag.STRING_IA5 -> decodeToString()
+            Tag.STRING_BMP -> decodeToString()
+            Tag.STRING_T61 -> decodeToString()
+            Tag.STRING_PRINTABLE -> decodeToString()
+            Tag.STRING_NUMERIC -> decodeToString()
+            Tag.STRING_VISIBLE -> decodeToString()
+            Tag.TIME_GENERALIZED -> decodeToInstant().toString()
+            Tag.TIME_UTC -> decodeToInstant().toString()
+            else -> content.toHexString(HexFormat.UpperCase)
+        }
+    }.getOrElse { "Non-complaint content: 0x" + content.toHexString(HexFormat.UpperCase) }
+
+
+    override fun prettyPrintContents(indent: Int) = " "+contentToString()
 
 
     override fun hashCode() = 31 * super.hashCode() + content.contentHashCode()
@@ -831,7 +903,7 @@ open class Asn1Primitive(
  * If you are expecting arbitrary bytes, [Asn1OctetString] is the correct type to look for.
  * (Your arbitrary bytes might inadvertently be valid ASN.1!)
  */
-interface Asn1OctetString {
+sealed interface Asn1OctetString {
 
     /**
      * Raw data contained in this ASN.1 primitive in its encoded form. Requires decoding to interpret it.
@@ -844,6 +916,7 @@ interface Asn1OctetString {
         /** Constructs a new ASN.1 OCTET STRING primitive containing these bytes */
         operator fun invoke(bytes: ByteArray) =
             Asn1PrimitiveOctetString(bytes)
+
         /** Constructs a new ASN.1 OCTET STRING primitive encapsulating these children */
         operator fun invoke(children: List<Asn1Element>) =
             Asn1EncapsulatingOctetString(children)